# docs/swagger.yaml
basePath: /
...
host: localhost:3000
info:
  contact:
    email: support@aygit.com
    name: API Destek
    url: http://www.aygit.com/destek
  description: Bu servis, NES Özel Entegratör üzerinden e-fatura, e-arşiv ve ürün
    yönetim süreçlerini yönetir.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Aygıt Muhasebe Entegrasyonu API
  version: "1.0"
paths:
...
  /api/v1/{type}/vouchers/{uuid}/savecompanyindocument:
    post:
      consumes:
      - application/json
      description: Firma olarak kaydet
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Save Company In Document
      tags:
      - Vouchers
  /api/v1/{type}/vouchers/{uuid}/usernotes:
    post:
      consumes:
      - application/json
      description: Kullanıcı notu ekler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Add User Note
      tags:
      - Vouchers
  /api/v1/{type}/vouchers/{uuid}/usernotes/{id}:
    delete:
      consumes:
      - application/json
      description: Kullanıcı notunu siler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete User Note
      tags:
      - Vouchers
    put:
      consumes:
      - application/json
      description: Kullanıcı notunu günceller
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update User Note
      tags:
      - Vouchers
  /api/v1/{type}/vouchers/bulk/{operation}:
    put:
      consumes:
      - application/json
      description: Yeni durum atar
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Operation
        in: path
        name: operation
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Bulk Operations
      tags:
      - Vouchers
  /api/v1/{type}/vouchers/cancel:
    post:
      consumes:
      - application/json
      description: Belgeyi iptal eder
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Cancel Voucher
      tags:
      - Vouchers
  /api/v1/{type}/vouchers/canceled/withdraw:
    post:
      consumes:
      - application/json
      description: İptal işlemini geri alır
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Withdraw Cancel Voucher
      tags:
      - Vouchers
  /api/v1/{type}/vouchers/drafts:
    delete:
      consumes:
      - application/json
      description: Taslak belgeleri silmek için bu uç kullanılablir
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete Drafts
      tags:
      - Vouchers
    get:
      description: Taslak makbuzları listeler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Draft Vouchers (Taslak Makbuzları Listeleme)
      tags:
      - Vouchers
  /api/v1/{type}/vouchers/email/send:
    post:
      consumes:
      - application/json
      description: Belgeyi mail olarak iletir
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Send Voucher Email
      tags:
      - Vouchers
  /api/v1/{type}/vouchers/export/{fileType}:
    post:
      consumes:
      - application/json
      description: Toplu aktar
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: File Type
        in: path
        name: fileType
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Export Vouchers
      tags:
      - Vouchers
  /api/v1/{type}/vouchers/tags:
    put:
      consumes:
      - application/json
      description: Etiket ekler/çıkarır
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Add/Remove Tags
      tags:
      - Vouchers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"